# Conflicts:
#	etomica-apps/src/main/java/etomica/spin/heisenberg/P2Spin.java
#	etomica-core/src/main/java/etomica/potential/PotentialCalculationPhiSumHeisenberg.java

+++ etomica-graph/src/main/java/etomica/graph/traversal/BitmapUtils.java
@@ -31,7 +31,7 @@ public class BitmapUtils {
   public static byte leftmostBit(int map) {
 
     for (byte i = 0; i < SZ_INT; i++) {
-      if ((map & bitOnMask((byte) i)) > 0) {
+      if ((map & bitOnMask(i)) > 0) {
         return i;
       }
     }
